+++ src/assos/assos.controller.ts
@@ -24,7 +24,7 @@ import AssosMemberCreateReqDto from './dto/req/assos-member-create.dto';
 import AssosMemberUpdateReqDto from './dto/req/assos-member-update.dto';
 import AssoMembershipResDto from './dto/res/assos-membership-res.dto';
 import UsersService from '../users/users.service';
-import { ConfigModule } from '../config/config.module';
+import { ConfigService } from '../config/config.service';
 import AssosUpdateReqDto from './dto/req/assos-update-req.dto';
 import { ImageMediaPreset } from '../prisma/types';
 import { LexicalModule } from '../lexical/lexical.module';
@@ -35,7 +35,7 @@ export class AssosController {
   constructor(
     readonly assosService: AssosService,
     readonly userService: UsersService,
-    readonly config: ConfigModule,
+    readonly config: ConfigService,
     readonly mediaService: ImageMediaService,
     readonly lexicalModule: LexicalModule,
   ) {}
