+++ src/auth/auth.controller.ts
@@ -17,7 +17,7 @@ import CreateApiKeyReqDto from './dto/req/create-api-key-req.dto';
 import ApplicationService from './application/application.service';
 import { Application } from './application/interfaces/application.interface';
 import AuthValidateReqDto from './dto/req/auth-validate-req.dto';
-import { ConfigModule } from '../config/config.module';
+import { ConfigService } from '../config/config.service';
 import AuthTokenResDto from './dto/res/auth-token-res.dto';
 import AuthRedirectionResDto from './dto/res/auth-redirection-res.dto';
 
@@ -28,7 +28,7 @@ export class AuthController {
     private authService: AuthService,
     private usersService: UsersService,
     private applicationService: ApplicationService,
-    private config: ConfigModule,
+    private config: ConfigService,
   ) {}
 
   @IsPublic()
