+++ src/config/config.module.ts
@@ -1,8 +1,9 @@
 import { Global, Module } from '@nestjs/common';
-import { ConfigModule as NestConfigModule, ConfigService as NestConfigService } from '@nestjs/config';
+import { ConfigModule as NestConfigModule } from '@nestjs/config';
+import { ConfigService } from './config.service';
 
-const isTestEnv = process.env.NODE_ENV === 'test';
-const isProdEnv = process.env.NODE_ENV === 'production';
+export const isTestEnv = process.env.NODE_ENV === 'test';
+export const isProdEnv = process.env.NODE_ENV === 'production';
 
 @Global()
 @Module({
@@ -14,90 +15,7 @@ const isProdEnv = process.env.NODE_ENV === 'production';
       envFilePath: isTestEnv ? '.env.test' : '.env.dev',
     }),
   ],
-  exports: [ConfigModule],
+  providers: [ConfigService],
+  exports: [ConfigService],
 })
-export class ConfigModule {
-  // BOTH DEV AND TEST ENVIRONMENTS
-  public readonly PAGINATION_PAGE_SIZE: number;
-  public readonly DATABASE_URL: string;
-  public readonly JWT_SECRET: string;
-  public readonly JWT_EXPIRES_IN: string;
-  public readonly SALT_ROUNDS: number;
-  public readonly CAS_URL: string;
-  public readonly CAS_SERVICE: string;
-  public readonly LDAP_URL: string;
-  public readonly LDAP_USER: string;
-  public readonly LDAP_PWD: string;
-  public readonly IS_PROD_ENV: boolean;
-  public readonly TIMETABLE_URL: string;
-  public readonly ANNAL_UPLOAD_DIR: string;
-  public readonly MEDIA_UPLOAD_DIR: string;
-  public readonly MEDIA_DETACHED_LIFESPAN: number;
-  public readonly ETUUTT_WEBSITE_APPLICATION_ID: string;
-  public readonly SMTP_HOST: string;
-  public readonly SMTP_PORT: number;
-  public readonly SMTP_USER: string;
-  public readonly SMTP_PASSWORD: string;
-  public readonly SMTP_REJECT_UNAUTHORIZED: boolean;
-  public readonly SMTP_SERVER_NAME: string;
-  public readonly SMTP_SENDING_NAME: string;
-  public readonly SMTP_SENDING_EMAIL: string;
-  public readonly WEEKLY_SEND_DAY: number;
-  public readonly WEEKLY_SEND_HOUR: number;
-  // DEV ENVIRONMENT ONLY
-
-  // TEST ENVIRONMENT ONLY
-  public readonly _FAKER_SEED: number;
-
-  constructor(config: NestConfigService) {
-    this.PAGINATION_PAGE_SIZE = Number(config.get('PAGINATION_PAGE_SIZE'));
-    this.DATABASE_URL = config.get('DATABASE_URL');
-    this.JWT_SECRET = config.get('JWT_SECRET');
-    this.JWT_EXPIRES_IN = config.get('JWT_EXPIRES_IN');
-    this.SALT_ROUNDS = Number(config.get('SALT_ROUNDS'));
-    this.CAS_URL = config.get('CAS_URL');
-    this.CAS_SERVICE = config.get('CAS_SERVICE');
-    this.LDAP_URL = config.get('LDAP_URL');
-    this.LDAP_USER = config.get('LDAP_USER');
-    this.LDAP_PWD = config.get('LDAP_PWD');
-    this.IS_PROD_ENV = isProdEnv;
-    this.TIMETABLE_URL = config.get<string>('TIMETABLE_URL');
-    this.ANNAL_UPLOAD_DIR = config.get<string>('ANNAL_UPLOAD_DIR');
-    if (this.ANNAL_UPLOAD_DIR.endsWith('/')) this.ANNAL_UPLOAD_DIR = this.ANNAL_UPLOAD_DIR.slice(0, -1);
-    this.MEDIA_UPLOAD_DIR = config.get<string>('MEDIA_UPLOAD_DIR');
-    if (this.MEDIA_UPLOAD_DIR.endsWith('/')) this.MEDIA_UPLOAD_DIR = this.MEDIA_UPLOAD_DIR.slice(0, -1);
-    this.MEDIA_DETACHED_LIFESPAN = Number(config.get('MEDIA_DETACHED_LIFESPAN'));
-    this.ETUUTT_WEBSITE_APPLICATION_ID = config.get('ETUUTT_WEBSITE_APPLICATION_ID');
-    this.SMTP_HOST = config.get<string>('SMTP_HOST');
-    this.SMTP_PORT = config.get<number>('SMTP_PORT');
-    this.SMTP_USER = config.get<string>('SMTP_USER');
-    this.SMTP_PASSWORD = config.get<string>('SMTP_PASSWORD');
-    this.SMTP_REJECT_UNAUTHORIZED = this.getBoolean(config.get('SMTP_REJECT_UNAUTHORIZED'));
-    this.SMTP_SERVER_NAME = config.get<string>('SMTP_SERVER_NAME');
-    this.SMTP_SENDING_NAME = config.get<string>('SMTP_SENDING_NAME');
-    this.SMTP_SENDING_EMAIL = config.get<string>('SMTP_SENDING_EMAIL');
-    this.WEEKLY_SEND_DAY = Number.parseInt(config.get('WEEKLY_SEND_DAY'));
-    this.WEEKLY_SEND_HOUR = Number.parseInt(config.get('WEEKLY_SEND_HOUR'));
-
-    this._FAKER_SEED = isTestEnv ? Number(config.get('FAKER_SEED')) : undefined;
-  }
-
-  get FAKER_SEED() {
-    if (!isTestEnv) throw new Error('FAKER_SEED is a test-environment-only environment variable');
-    return this._FAKER_SEED;
-  }
-
-  get<T extends keyof ConfigModule>(key: T): ConfigModule[T] {
-    return this[key];
-  }
-
-  private getBoolean(value: string): boolean {
-    if (value == 'true') {
-      return true;
-    } else if (value == 'false') {
-      return false;
-    } else {
-      throw new Error('Environment variable SMTP_REJECT_UNAUTHORIZED must be "true" or "false"');
-    }
-  }
-}
+export class ConfigModule {}
