+++ src/ue/ue.service.ts
@@ -5,7 +5,7 @@ import { UeRateReqDto } from './dto/req/ue-rate-req.dto';
 import { Ue } from './interfaces/ue.interface';
 import { Criterion } from './interfaces/criterion.interface';
 import { UeRating } from './interfaces/rate.interface';
-import { ConfigModule } from '../config/config.module';
+import { ConfigService } from '../config/config.service';
 import { Language, Prisma } from '../prisma/types';
 import { SemesterService } from '../semester/semester.service';
 
@@ -13,7 +13,7 @@ import { SemesterService } from '../semester/semester.service';
 export class UeService {
   constructor(
     readonly prisma: PrismaService,
-    readonly config: ConfigModule,
+    readonly config: ConfigService,
     readonly semesterService: SemesterService,
   ) {}
