+++ test/e2e/app.e2e-spec.ts
@@ -1,3 +1,10 @@
+jest.mock('@nestjs-modules/mailer/dist/adapters/ejs.adapter', () => ({
+  EjsAdapter: jest.fn().mockImplementation(() => ({
+    __esModule: true,
+    compile: jest.fn(),
+  })),
+}));
+
 import '../declarations';
 import '../../src/std.type';
 import * as testUtils from '../utils/test_utils';
@@ -13,7 +20,7 @@ import UeE2ESpec from './ue';
 import { AppValidationPipe } from '../../src/app.pipe';
 import * as cas from '../external_services/cas';
 import * as timetableProvider from '../external_services/timetable';
-import { ConfigModule } from '../../src/config/config.module';
+import { ConfigService } from '../../src/config/config.service';
 import AssoE2ESpec from './assos';
 import MediaE2ESpec from './media';
 
@@ -41,12 +48,12 @@ describe('EtuUTT API e2e testing', () => {
         process.env.API_PREFIX.endsWith('/') ? '' : '/'
       }v1`,
     );
-    cas.enable(app.get(ConfigModule));
+    cas.enable(app.get(ConfigService));
     timetableProvider.enable('https://monedt.utt.fr/calendrier');
   });
 
-  afterAll(() => {
-    app.close();
+  afterAll(async () => {
+    await app.close();
   });
 
   AuthE2ESpec(() => app);
