+++ test/e2e/assos/get-weekly-info.e2e-spec.ts
@@ -2,7 +2,7 @@ import { e2eSuite } from '../../utils/test_utils';
 import { createUser } from '../../utils/fakedb';
 import * as pactum from 'pactum';
 import { ERROR_CODE } from '../../../src/exceptions';
-import { ConfigModule } from '../../../src/config/config.module';
+import { ConfigService } from '../../../src/config/config.service';
 
 const GetWeeklyInfoE2ESpec = e2eSuite('GET /assos/weekly/info', (app) => {
   const user = createUser(app);
@@ -15,7 +15,7 @@ const GetWeeklyInfoE2ESpec = e2eSuite('GET /assos/weekly/info', (app) => {
       .spec()
       .withBearerToken(user.token)
       .get('/assos/weekly/info')
-      .expectJson({ sendDay: app().get(ConfigModule).WEEKLY_SEND_DAY, sendHour: app().get(ConfigModule).WEEKLY_SEND_HOUR }));
+      .expectJson({ sendDay: app().get(ConfigService).WEEKLY_SEND_DAY, sendHour: app().get(ConfigService).WEEKLY_SEND_HOUR }));
 });
 
 export default GetWeeklyInfoE2ESpec;
