+++ test/e2e/ue/annals/upload-annal.e2e-spec.ts
@@ -11,7 +11,7 @@ import {
 } from '../../../utils/fakedb';
 import { JsonLike, e2eSuite } from '../../../utils/test_utils';
 import { ERROR_CODE } from '../../../../src/exceptions';
-import { ConfigModule } from '../../../../src/config/config.module';
+import { ConfigService } from '../../../../src/config/config.service';
 import { CommentStatus } from 'src/ue/comments/interfaces/comment.interface';
 import { PermissionManager, pick } from '../../../../src/utils';
 import { mkdirSync, rmSync } from 'fs';
@@ -116,11 +116,11 @@ const PostAnnal = e2eSuite('POST-PUT /ue/annals', (app) => {
 
   describe('should create the annal', () => {
     beforeAll(() => {
-      mkdirSync(app().get(ConfigModule).ANNAL_UPLOAD_DIR, { recursive: true });
+      mkdirSync(app().get(ConfigService).ANNAL_UPLOAD_DIR, { recursive: true });
     });
 
     afterAll(() => {
-      rmSync(app().get(ConfigModule).ANNAL_UPLOAD_DIR.split('/')[0], { recursive: true });
+      rmSync(app().get(ConfigService).ANNAL_UPLOAD_DIR.split('/')[0], { recursive: true });
     });
 
     const testFunction =
