+++ test/external_services/ldap.ts
@@ -1,7 +1,6 @@
 import { LdapServerMock, LdapUser } from 'ldap-server-mock';
 
 export function mockLdapServer(list: LdapUser[]) {
-  //console.log("loading server ", process.env.LDAP_URL)
   const ldapServer = new LdapServerMock(
     list,
     {
