+++ test/unit/app.spec.ts
@@ -1,3 +1,9 @@
+jest.mock('@nestjs-modules/mailer/dist/adapters/ejs.adapter', () => ({
+  EjsAdapter: jest.fn().mockImplementation(() => ({
+    compile: jest.fn(),
+  })),
+}));
+
 import TimetableServiceUnitSpec from './timetable/timetable.service.spec';
 import LexicalValidationUnitSpec from './lexical/lexical-validation.spec';
 import LexicalGenerationUnitSpec from './lexical/lexical-generation.spec';
@@ -10,6 +16,9 @@ describe('EtuUTT API unit testing', () => {
   beforeAll(async () => {
     app = await Test.createTestingModule({ imports: [AppModule] }).compile();
   });
+  afterAll(async () => {
+    await app.close();
+  })
   TimetableServiceUnitSpec(() => app);
   LexicalValidationUnitSpec(() => app);
   LexicalGenerationUnitSpec(() => app);
